move game methods from utils to game.go
